+++ src/main/java/com/example/bigjavahomework/services/implementation/CountriesImplementation.java
@@ -1,11 +1,18 @@
 package com.example.bigjavahomework.services.implementation;
 
+import jakarta.persistence.EntityNotFoundException;
+import lombok.RequiredArgsConstructor;
+import com.example.bigjavahomework.resources.CountriesResource;
 import com.example.bigjavahomework.entityes.Countries;
 import com.example.bigjavahomework.repositoryes.CountriesRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
 import com.example.bigjavahomework.services.CountriesService;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.stereotype.Service;
+
 import java.util.List;
+import java.util.Optional;
+
+import static com.example.bigjavahomework.mappers.CountriesMapper.COUNTRIES_MAPPER;
 
 @Service
 @RequiredArgsConstructor
@@ -13,34 +20,47 @@ public class CountriesImplementation implements CountriesService {
     private final CountriesRepository countriesRepository;
 
     @Override
-    public List<Countries> allCountries() {
-        return countriesRepository.findAll();
+    public List<CountriesResource> getAll() {
+        return COUNTRIES_MAPPER.toCountriesResources(countriesRepository.findAll());
     }
 
     @Override
-    public Countries save(Countries country) {
-        return countriesRepository.save(country);
+    public Optional<CountriesResource> getByCode(String code) {
+        return countriesRepository.findById(code).map(COUNTRIES_MAPPER::toCountriesResource);
     }
 
     @Override
-    public Countries update(Countries country) {
-        Countries existingCountry = countriesRepository.findById(country.getCode()).orElse(null);
-        if (existingCountry != null) {
-            existingCountry.setName(country.getName());
-            existingCountry.setLocation(country.getLocation());
-            existingCountry.setAllowed(country.getAllowed());
-            return countriesRepository.save(existingCountry);
+    public CountriesResource create(CountriesResource countriesResource) {
+        try {
+            Countries countries = countriesRepository.save(COUNTRIES_MAPPER.fromCountriesResource(countriesResource));
+            countriesResource.setCode(countries.getCode());
+            return countriesResource;
+        } catch (DataIntegrityViolationException e) {
+            throw new DataIntegrityViolationException("Country with code " + countriesResource.getCode() + " already exists.");
         }
-        return countriesRepository.save(country);
+    }
+
+    @Override
+    public CountriesResource update(CountriesResource countriesResource, String code) {
+        try {
+            Countries countries = countriesRepository.getReferenceById(code);
+
+            countries.setName(countriesResource.getName());
+            countries.setLocation(countriesResource.getLocation());
+            countries.setAllowed(countriesResource.getAllowed());
+
+            return COUNTRIES_MAPPER.toCountriesResource(countriesRepository.save(countries));
+        } catch (EntityNotFoundException e) {
+            throw new EntityNotFoundException("Unable to find country with code " + code + ".");
+        }
     }
 
     @Override
     public void deleteById(String code) {
-        countriesRepository.deleteById(code);
-    }
-
-    @Override
-    public Countries findById(String code) {
-        return countriesRepository.findById(code).orElse(null);
+        if (countriesRepository.existsById(code)) {
+            countriesRepository.deleteById(code);
+        } else {
+            throw new EntityNotFoundException("Unable to find country with code " + code + ".");
+        }
     }
 }
